media: refactoring to use API Version `2022-08-01` for `StreamingEndpoints`
